requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Session==0.8.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
stripe==7.8.0
reportlab==4.0.7
Werkzeug==3.0.1
email-validator==2.1.0
openai==2.6.0
psycopg2-binary==2.9.9
redis==5.0.1
gunicorn==21.2.0
